Adds English traineddata as APK asset.
Adds a default English trained data set to the APK. The API requires
a File object to do processing, so the assets dir will only serve
as a starting point. Plan is to unpack to persistent external storage
in the long term. Will need a system to check whether to recreate
the persistent storage with default data when missing.